Relocate util functions to util folder
